lua REPL fix
this is a weird one.

i was getting unexpected crashes when running certain chunks of lua code - namely, an undecorated call to a registered C function with no return values (like `report_engines()`.)

traced this to `add_return()` in lua_eval.c. this function is copied near-verbatim from the lua source. it prepends the string to be compiled with "return" (for returning expression results), appends a semicolon, and calls `luaL_loadbuffer` with size `strlen(buffer)`.

this seems wrong to me - the semicolon is unnecessary, and not allowing for the terminating null character is a classic case for a buffer overflow in C. but i don't totally understand the internals of `luaL_loadbuffer` - apparently it calls `luac` (the lua compiler) and luac is somehow supposed to determine whether the buffer is a string or a compiled binary chunk.

in the crash case, after loading the buffer and calling the resulting chunk, the lua stack was full of garbage and any attempt to access it would segfault deep within the lua dylib. looks like buffer overflow to me.

well, i just removed the semicolon and added an extra byte to the size argument for the terminating null, and it seems to work. the mystery as ever is why it ever worked before (or indeed, why it works in the lua standalone) - it could just be the vagaries of memory layout, but i can't really know without looking at the luac sources.

will keep an eye on this and maybe post a SO question.