Append port index only for multi-port devices
The port index was being appended for devices which only have
one port and this doesn't necessarily make sense. This is being
removed by passing a bool indicating whether the port being
referenced is the sole port of the MIDI device, and only in this
case will the index be appended.