Use device hostname for wifi hotspot ssid
Currently the hotspot ssid / password are hard coded to "norns" / "nnnnnnnn"

In a situation with multiple devices, having the script pick up the device hostname instead would be nice. Mostly applicable to someone with norns + shield, or shield + shield, etc.

I've added a variable `hotspot_password` but I'm not sure if there's a better place to establish that. For the moment it's local to `wifi.lua`. (also not sure if there's a preferred variable naming scheme I should use).