skip virtual midi interface when auto-assigning vports
the virtual midi interface is a specialized device which
is by default not connected to anything. this change avoids
auto-assigning the virtual device to a vport. one must explicitly
assign the virtual device to a vport in order to use it. once
explicitly assigned the configuration is retained.

the change preserves existing behavior of assigning attached
physical devices to the next available vport when no explict
assignment has been made.