allow parameter mapping to be disabled
some script parameters may not be suitable for live
manipulation via mapped midi controller. this change
adds an optional `allow_pmap` property to `Control`,
`Option`, and `Number` parameter types.

the `get_allow_pmap(...)` method on `ParamSet` will
return the property if set on the param otherwise it will
return `true` in order to retain existing behavior.

the menu code was changed to prevent access to the
MAPEDIT menu unless the `get_allow_pmap(...)` is `true`.