minor build improvements (#1281)
* fix maiden-repl dependency checking

* edge.sh: fail fast, limit build parallelism

assuming the target audience for `edge.sh` is a less technical
person testing unreleased features it makes sense to exit as
soon as one of the commands fails (such as `waf configure`) instead
of continuing on to produce a bad build or errors which obscure
the initial problem.

limiting parallelism is a safety. building more than two of the
c++ components in parallel with the matron/crone/supercollider running
will somewhat often run out of memory and fail. limiting the build
to two jobs leaves ~100Mb of headroom during a clean build