fix: always push four quads to libmonome
when grid rotation is nonzero, device_monome.c is writing to quads
that end up never being pushed to the device (quad 3,
specifically). i'm not familiar with libmonome, and i have yet to test
this on my norns, but this is the solution that i used in
ryleelyman/seamstress#113 to get grid rotation working as expected.

closes #1767