feat: add `clock.link.get_number_of_peers()` (#1825)
* feat(matron): maintain knowledge of number of Link peers

this commit adds `number_of_peers` to `clock_link_shared_data_t` and
keeps this data in sync with the Link session. in the header, a new
function to return this number of peers is added.

this commit does not yet expose this functionality to Lua.

* feat(matron): add `_norns.clock_link_get_number_of_peers()`

this commit adds (provided `HAVE_ABLETON_LINK` is true) a new Lua
function which will return the number of peers connected via Ableton
Link. When Link functionality is disabled, this function returns 0.

this commit does not add any UI changes.

* fix: get number of peers on demand

rather than managing the number of link peers automatically, the Lua
instance will now query the Link library directly.

* fix: remove unnecessary `number_of_peers` field

* fix: remove erroneous number_of_peers access

* fix: norns should compile and run without errors

* nit: link_instance -> clock_link

* feat: add `clock.link.get_number_of_peers()` to Lua